ref: rename dirs
