+++ examples/erc721-metadata/tests/abi/mod.rs
