+++ examples/erc721/tests/mock/mod.rs
@@ -1,0 +1,1 @@
+pub mod receiver;
