--- examples/erc721/tests/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod abi;
-pub mod mock_receiver;
