--- examples/ownable/tests/utils/mod.rs
@@ -1,1 +1,0 @@
-pub mod abi;
